start renaming the boarChecker class

+++ src/game/TTElement.java
@@ -2,7 +2,7 @@ package game;
 
 public class TTElement {
     public final long primaryHash;
-    public final int value;
+    public final byte value;
     public final byte depth;
     //flag: 0 exact value, 1: lower bound, 2:upper bound
     public final byte flag;
@@ -11,7 +11,7 @@ public class TTElement {
     public TTElement(long primaryHash, int flag, int value, int depth, Move move){
         this.primaryHash = primaryHash;
         this.flag = (byte)flag;
-        this.value = value;
+        this.value = (byte)value;
         this.depth = (byte)depth;
         this.move = move;
     }
